pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "n64tex"
version = "0.1.0"
requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "numpy",
    "Pillow"
]

# [tool.setuptools]
# packages = ["dk64_lib", "dk64_lib.components"]

# [tool.setuptools.packages]
# find = {}

# [tool.setuptools.packages.find]
# where = ["dk64_lib", "dk64_lib.*"]

# [tool.setuptools.package-dir]
# dk64_lib = "dk64_lib"
# "dk64_lib.components" = "components"